split up authorizer.go mega-file
